Move updatePod to statefulset_test
